moved db connection to separate file for ease of maintenance; test using exit()
